app/requirements.txt
numpy>=1.24
onnxruntime>=1.17.0
open3d>=0.18.0
plotly>=5.20.0
pyyaml>=6.0
rosbags>=0.10.0
streamlit>=1.34.0
